Fix some issues

+++ ExoabEngine/src/utils/perlin_noise.hpp
@@ -2,5 +2,5 @@
 #include <stdint.h>
 
 namespace Utils {
-	void perlin(uint32_t width, uint32_t height, uint32_t seed, float frequency, float octaves, uint8_t* output_buffer);
+	void perlin(uint32_t width, uint32_t height, uint32_t seed, float frequency, float octaves, float* output_buffer);
 }
